Hack in a working else statement
This appears to be errata in the series itself. Note that Crenshaw's
block statement treats 'o' as a special other character, when all
non token chars should be other. If you fix this, then the call to block
before the check inside doIf for 'l' will swallow l and treat it like
an other character.